refactor: simplify folder structure
