+++ reflekt/cli.py
@@ -16,15 +16,15 @@ from loguru import logger
 from packaging.version import InvalidVersion
 from packaging.version import parse as parse_version
 
+from reflekt import constants
+from reflekt.api_handler import ReflektApiHandler
 from reflekt.avo.plan import AvoPlan
+from reflekt.config import ReflektConfig
+from reflekt.loader import ReflektLoader
 from reflekt.logger import logger_config
-from reflekt.reflekt import constants
-from reflekt.reflekt.api_handler import ReflektApiHandler
-from reflekt.reflekt.config import ReflektConfig
-from reflekt.reflekt.loader import ReflektLoader
-from reflekt.reflekt.project import ReflektProject
-from reflekt.reflekt.transformer import ReflektTransformer
+from reflekt.project import ReflektProject
 from reflekt.segment.plan import SegmentPlan
+from reflekt.transformer import ReflektTransformer
 
 # from ..reflekt import __version__
