+++ reflekt/property.py
@@ -8,10 +8,10 @@ import re
 
 from cerberus import Validator
 
-from reflekt.reflekt.casing import CAMEL_CASE_RE, SNAKE_CASE_RE, TITLE_CASE_RE
-from reflekt.reflekt.errors import ReflektValidationError
-from reflekt.reflekt.project import ReflektProject
-from reflekt.reflekt.schema import (
+from reflekt.casing import CAMEL_CASE_RE, SNAKE_CASE_RE, TITLE_CASE_RE
+from reflekt.errors import ReflektValidationError
+from reflekt.project import ReflektProject
+from reflekt.schema import (
     reflekt_item_schema,
     reflekt_nested_property_schema,
     reflekt_property_schema,
