--- reflekt/reflekt/__init__.py
@@ -1,3 +1,0 @@
-# SPDX-FileCopyrightText: 2022 Gregory Clunies <greg@reflekt-ci.com>
-#
-# SPDX-License-Identifier: Apache-2.0
