+++ reflekt/segment/api.py
@@ -8,7 +8,7 @@ from typing import Optional
 import requests
 from requests import Response
 
-from reflekt.reflekt.config import ReflektConfig
+from reflekt.config import ReflektConfig
 from reflekt.segment.errors import SegmentApiError
