+++ reflekt/segment/plan.py
@@ -13,7 +13,7 @@ import yaml
 from inflection import dasherize, underscore
 from loguru import logger
 
-from reflekt.reflekt.dumper import ReflektYamlDumper
+from reflekt.dumper import ReflektYamlDumper
 from reflekt.segment.parser import parse_segment_event, parse_segment_property
