+++ tests/test_reflekt_loader.py
@@ -5,10 +5,10 @@
 import pytest
 import yaml
 
-from reflekt.reflekt.errors import ReflektValidationError
-from reflekt.reflekt.event import ReflektEvent
-from reflekt.reflekt.loader import ReflektLoader
-from reflekt.reflekt.plan import ReflektPlan
+from reflekt.errors import ReflektValidationError
+from reflekt.event import ReflektEvent
+from reflekt.loader import ReflektLoader
+from reflekt.plan import ReflektPlan
 from tests.fixtures import (
     REFLEKT_EVENT,
     REFLEKT_EVENT_BAD,
