+++ tests/test_reflekt_property.py
@@ -5,8 +5,8 @@
 import pytest
 import yaml
 
-from reflekt.reflekt.errors import ReflektValidationError
-from reflekt.reflekt.property import ReflektProperty
+from reflekt.errors import ReflektValidationError
+from reflekt.property import ReflektProperty
 from tests.fixtures import REFLEKT_PROPERTY, REFLEKT_PROPERTY_BAD
