+++ tests/test_reflekt_transformer.py
@@ -3,8 +3,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
 
-from reflekt.reflekt.loader import ReflektLoader
-from reflekt.reflekt.transformer import ReflektTransformer
+from reflekt.loader import ReflektLoader
+from reflekt.transformer import ReflektTransformer
 from tests.fixtures import _build_reflekt_plan_dir
